python/weather-forecast-api/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
azure-keyvault-secrets==4.8.0
azure-identity==1.15.0
